+++ web/template/partial/stream/chat/emojipromt.gohtml
@@ -1,8 +1,12 @@
 {{define "emojipromt"}}
-    <div x-show="emojiSuggestions.length>0" class="text-left overflow-hidden whitespace-nowrap px-2">
-        <template x-for="emoji in emojiSuggestions">
-                <span x-text="emoji.v"
-                      class="cursor-pointer mr-1 p-1 bg-gray-100 dark:bg-gray-700 shadow rounded "></span>
+    <div x-show="emojiSuggestions.length>0" class="text-left overflow-hidden whitespace-nowrap px-2 flex flex-row">
+        <template x-for="(emoji, i) in emojiSuggestions">
+                <button :title="`:${emoji.k[0]}:`" {{/* e.g. ':eyes:' */}}
+                        x-text="emoji.v"
+                        x-on:mouseenter="emojiIndex=i" {{/* Select emoji on hover */}}
+                        :class="emojiIndex===i && 'border-sky-600 border-2'"
+                        @click="$dispatch('emojiselectiontriggered')" {{/* notify chatpromts to insert emoji */}}
+                        class="cursor-pointer dark:bg-gray-700 bg-gray-200 mr-1 p-1 shadow rounded"></button>
         </template>
     </div>
 {{end}}
