+++ web/ts/watch.ts
@@ -1,19 +1,6 @@
-import { Emoji, findEmojisForInput } from "./chat";
+import { showNewMessageIndicator, scrollChat, shouldScroll, showDisconnectedMsg, hideDisconnectedMsg } from "./chat";
 
-export async function getEmojisForMessage(message: string, cursorPos: number): Promise<Emoji[]> {
-    let currentWord = "";
-    // backtrack from cursor position to get the word currently being typed
-    while (cursorPos >= 0 && message.charAt(cursorPos) !== " ") {
-        currentWord = message.charAt(cursorPos--) + currentWord;
-    }
-    if (!currentWord.startsWith(":") || currentWord.length < 2) {
-        return [];
-    }
-    currentWord = currentWord.substring(1);
-    return findEmojisForInput(currentWord);
-}
-
-let chatInput;
+let chatInput: HTMLInputElement;
 
 export class Watch {
     constructor() {
@@ -31,7 +18,6 @@ export class Watch {
 
 let ws: WebSocket;
 let retryInt = 5000; //retry connecting to websocket after this timeout
-let orderByLikes = false; // sorting by likes or by time
 
 const scrollDelay = 100; // delay before scrolling to bottom to make sure chat is rendered
 const pageloaded = new Date();
@@ -60,23 +46,6 @@ export function deleteMessage(id: number) {
     );
 }
 
-export function sortMessages(messages): void {
-    messages.sort((m1, m2) => {
-        if (orderByLikes) {
-            if (m1.likes === m2.likes) {
-                return m2.id - m1.id; // same amount of likes -> newer messages up
-            }
-            return m2.likes - m1.likes; // more likes -> up
-        } else {
-            return m1.ID < m2.ID ? -1 : 1; // newest messages last
-        }
-    });
-}
-
-export function setOrder(obl: boolean) {
-    orderByLikes = obl;
-}
-
 export function initChatScrollListener() {
     const chatBox = document.getElementById("chatBox") as HTMLDivElement;
     if (!chatBox) {
@@ -87,32 +56,6 @@ export function initChatScrollListener() {
             window.dispatchEvent(new CustomEvent("messageindicator", { detail: { show: false } }));
         }
     });
-}
-
-export function shouldScroll(): boolean {
-    if (orderByLikes) {
-        return false; // only scroll if sorting by time
-    }
-    const c = document.getElementById("chatBox");
-    return c.scrollHeight - c.scrollTop <= c.offsetHeight;
-}
-
-function showNewMessageIndicator() {
-    window.dispatchEvent(new CustomEvent("messageindicator", { detail: { show: true } }));
-}
-
-export function scrollChat() {
-    if (orderByLikes) {
-        return; // only scroll if sorting by time
-    }
-    const c = document.getElementById("chatBox");
-    c.scrollTop = c.scrollHeight;
-}
-
-export function scrollToLatestMessage() {
-    const c = document.getElementById("chatBox");
-    c.scrollTo({ top: c.scrollHeight, behavior: "smooth" });
-    window.dispatchEvent(new CustomEvent("messageindicator", { detail: { show: false } }));
 }
 
 export function startWebsocket() {
@@ -217,16 +160,4 @@ export function sendMessage(message: string, anonymous: boolean, replyTo: number
             replyTo: replyTo,
         }),
     );
-}
-
-export function showDisconnectedMsg() {
-    if (document.getElementById("disconnectMsg") !== null) {
-        document.getElementById("disconnectMsg").classList.remove("hidden");
-    }
-}
-
-export function hideDisconnectedMsg() {
-    if (document.getElementById("disconnectMsg") !== null) {
-        document.getElementById("disconnectMsg").classList.add("hidden");
-    }
 }
